requirements.txt
aiohttp
tenacity
ujson
python-dateutil
beautifulsoup4
tqdm
